#12 winport branch merged + applying changes to linux, some tests need debug

+++ tcejdb/bson.c
@@ -1274,7 +1274,7 @@ void bson_numstr(char *str, int64_t i) {
     if (i >= 0 && i < 1000)
         memcpy(str, bson_numstrs[i], 4);
     else
-        sprintf(str, "%" PRIdMAX "", (long long int) i);
+        sprintf(str, "%" PRIdMAX "", (int64_t) i);
 }
 
 #pragma GCC diagnostic push
@@ -1285,7 +1285,7 @@ int bson_numstrn(char *str, int maxbuf, int64_t i) {
         memcpy(str, bson_numstrs[i], 4);
         return strlen(bson_numstrs[i]);
     } else {
-        return snprintf(str, maxbuf, "%" PRIdMAX "", (long long int) i);
+        return snprintf(str, maxbuf, "%" PRIdMAX "", (int64_t) i);
     }
 }
 #pragma GCC diagnostic pop
