+++ tcejdb/tcbmgr.c
@@ -662,18 +662,18 @@ static int procinform(const char *path, int omode){
   printf("\n");
   printf("max leaf member: %d\n", tcbdblmemb(bdb));
   printf("max node member: %d\n", tcbdbnmemb(bdb));
-  printf("leaf number: %" PRIuMAX "\n", (unsigned long long)tcbdblnum(bdb));
-  printf("node number: %" PRIuMAX "\n", (unsigned long long)tcbdbnnum(bdb));
-  printf("bucket number: %" PRIuMAX "\n", (unsigned long long)tcbdbbnum(bdb));
+  printf("leaf number: %" PRIuMAX "\n", (uint64_t)tcbdblnum(bdb));
+  printf("node number: %" PRIuMAX "\n", (uint64_t)tcbdbnnum(bdb));
+  printf("bucket number: %" PRIuMAX "\n", (uint64_t)tcbdbbnum(bdb));
 
 #ifndef NDEBUG
   if(bdb->hdb->cnt_writerec >= 0)
-    printf("used bucket number: %" PRIdMAX "\n", (long long)tcbdbbnumused(bdb));
-#endif  
+    printf("used bucket number: %" PRIdMAX "\n", (int64_t) tcbdbbnumused(bdb));
+#endif
 
   printf("alignment: %u\n", tcbdbalign(bdb));
   printf("free block pool: %u\n", tcbdbfbpmax(bdb));
-  printf("inode number: %" PRIdMAX "\n", (long long)tcbdbinode(bdb));
+  printf("inode number: %" PRIdMAX "\n", (int64_t) tcbdbinode(bdb));
   char date[48];
   tcdatestrwww(tcbdbmtime(bdb), INT_MAX, date);
   printf("modified time: %s\n", date);
@@ -685,8 +685,8 @@ static int procinform(const char *path, int omode){
   if(opts & BDBTTCBS) printf(" tcbs");
   if(opts & BDBTEXCODEC) printf(" excodec");
   printf("\n");
-  printf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  printf("file size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  printf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  printf("file size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   if(!tcbdbclose(bdb)){
     if(!err) printerr(bdb);
     err = true;
