+++ tcejdb/tcbmttest.c
@@ -203,34 +203,34 @@ static void mprint(TCBDB *bdb){
   if(bdb->hdb->cnt_writerec < 0) return;
   iprintf("max leaf member: %d\n", tcbdblmemb(bdb));
   iprintf("max node member: %d\n", tcbdbnmemb(bdb));
-  iprintf("leaf number: %" PRIdMAX "\n", (long long)tcbdblnum(bdb));
-  iprintf("node number: %" PRIdMAX "\n", (long long)tcbdbnnum(bdb));
-  iprintf("bucket number: %" PRIdMAX "\n", (long long)tcbdbbnum(bdb));
-  iprintf("used bucket number: %" PRIdMAX "\n", (long long)tcbdbbnumused(bdb));
-  iprintf("cnt_saveleaf: %" PRIdMAX "\n", (long long)bdb->cnt_saveleaf);
-  iprintf("cnt_loadleaf: %" PRIdMAX "\n", (long long)bdb->cnt_loadleaf);
-  iprintf("cnt_killleaf: %" PRIdMAX "\n", (long long)bdb->cnt_killleaf);
-  iprintf("cnt_adjleafc: %" PRIdMAX "\n", (long long)bdb->cnt_adjleafc);
-  iprintf("cnt_savenode: %" PRIdMAX "\n", (long long)bdb->cnt_savenode);
-  iprintf("cnt_loadnode: %" PRIdMAX "\n", (long long)bdb->cnt_loadnode);
-  iprintf("cnt_adjnodec: %" PRIdMAX "\n", (long long)bdb->cnt_adjnodec);
-  iprintf("cnt_writerec: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_writerec);
-  iprintf("cnt_reuserec: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_reuserec);
-  iprintf("cnt_moverec: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_moverec);
-  iprintf("cnt_readrec: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_readrec);
-  iprintf("cnt_searchfbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_searchfbp);
-  iprintf("cnt_insertfbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_insertfbp);
-  iprintf("cnt_splicefbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_splicefbp);
-  iprintf("cnt_dividefbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_dividefbp);
-  iprintf("cnt_mergefbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_mergefbp);
-  iprintf("cnt_reducefbp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_reducefbp);
-  iprintf("cnt_appenddrp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_appenddrp);
-  iprintf("cnt_deferdrp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_deferdrp);
-  iprintf("cnt_flushdrp: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_flushdrp);
-  iprintf("cnt_adjrecc: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_adjrecc);
-  iprintf("cnt_defrag: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_defrag);
-  iprintf("cnt_shiftrec: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_shiftrec);
-  iprintf("cnt_trunc: %" PRIdMAX "\n", (long long)bdb->hdb->cnt_trunc);
+  iprintf("leaf number: %" PRIdMAX "\n", (int64_t)tcbdblnum(bdb));
+  iprintf("node number: %" PRIdMAX "\n", (int64_t)tcbdbnnum(bdb));
+  iprintf("bucket number: %" PRIdMAX "\n", (int64_t)tcbdbbnum(bdb));
+  iprintf("used bucket number: %" PRIdMAX "\n", (int64_t)tcbdbbnumused(bdb));
+  iprintf("cnt_saveleaf: %" PRIdMAX "\n", (int64_t)bdb->cnt_saveleaf);
+  iprintf("cnt_loadleaf: %" PRIdMAX "\n", (int64_t)bdb->cnt_loadleaf);
+  iprintf("cnt_killleaf: %" PRIdMAX "\n", (int64_t)bdb->cnt_killleaf);
+  iprintf("cnt_adjleafc: %" PRIdMAX "\n", (int64_t)bdb->cnt_adjleafc);
+  iprintf("cnt_savenode: %" PRIdMAX "\n", (int64_t)bdb->cnt_savenode);
+  iprintf("cnt_loadnode: %" PRIdMAX "\n", (int64_t)bdb->cnt_loadnode);
+  iprintf("cnt_adjnodec: %" PRIdMAX "\n", (int64_t)bdb->cnt_adjnodec);
+  iprintf("cnt_writerec: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_writerec);
+  iprintf("cnt_reuserec: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_reuserec);
+  iprintf("cnt_moverec: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_moverec);
+  iprintf("cnt_readrec: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_readrec);
+  iprintf("cnt_searchfbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_searchfbp);
+  iprintf("cnt_insertfbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_insertfbp);
+  iprintf("cnt_splicefbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_splicefbp);
+  iprintf("cnt_dividefbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_dividefbp);
+  iprintf("cnt_mergefbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_mergefbp);
+  iprintf("cnt_reducefbp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_reducefbp);
+  iprintf("cnt_appenddrp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_appenddrp);
+  iprintf("cnt_deferdrp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_deferdrp);
+  iprintf("cnt_flushdrp: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_flushdrp);
+  iprintf("cnt_adjrecc: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_adjrecc);
+  iprintf("cnt_defrag: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_defrag);
+  iprintf("cnt_shiftrec: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_shiftrec);
+  iprintf("cnt_trunc: %" PRIdMAX "\n", (int64_t)bdb->hdb->cnt_trunc);
 }
 
 
@@ -710,8 +710,8 @@ static int procwrite(const char *path, int tnum, int rnum, int lmemb, int nmemb,
       }
     }
   }
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
@@ -788,8 +788,8 @@ static int procread(const char *path, int tnum, int xmsiz, int dfunit, int omode
       }
     }
   }
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
@@ -863,8 +863,8 @@ static int procremove(const char *path, int tnum, int xmsiz, int dfunit, int omo
       }
     }
   }
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
@@ -975,8 +975,8 @@ static int procwicked(const char *path, int tnum, int rnum, int opts, int omode,
     if(rnum % 50 > 0) iprintf(" (%08d)\n", rnum);
   }
   tcmapdel(map);
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
@@ -1060,8 +1060,8 @@ static int proctypical(const char *path, int tnum, int rnum, int lmemb, int nmem
       }
     }
   }
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
@@ -1139,8 +1139,8 @@ static int procrace(const char *path, int tnum, int rnum, int lmemb, int nmemb,
       }
     }
   }
-  iprintf("record number: %" PRIuMAX "\n", (unsigned long long)tcbdbrnum(bdb));
-  iprintf("size: %" PRIuMAX "\n", (unsigned long long)tcbdbfsiz(bdb));
+  iprintf("record number: %" PRIuMAX "\n", (uint64_t)tcbdbrnum(bdb));
+  iprintf("size: %" PRIuMAX "\n", (uint64_t)tcbdbfsiz(bdb));
   mprint(bdb);
   sysprint();
   if(!tcbdbclose(bdb)){
